create plot dir, refactor imports, separte logging from plotting. closes #117
