+++ retinal_rl/rl/analyze.py
@@ -5,9 +5,9 @@ import torch
 
 from retinal_rl.analysis import default as default_ana
 from retinal_rl.analysis import receptive_fields
-from retinal_rl.analysis.plot import FigureLogger
 from retinal_rl.models.brain import Brain
 from retinal_rl.models.objective import ContextT, Objective
+from retinal_rl.plot.log import FigureLogger
 
 ### Infrastructure ###
